[*] timed_event -> event. event.cpp, timer.cpp
