+++ conanfile.py
@@ -9,6 +9,7 @@ class CoLibConan(ConanFile):
     description = "experimental asynchronous C++20 framework that feels like std library"
     topics = ("c++20", "coroutines", "asynchronous programming")
     generators = "cmake"
+    settings = "os", "compiler", "build_type", "arch"
     exports_sources = "include/*", "CMakeLists.txt", "tests/*", "cmake/*"
     requires = "libuv/1.40.0", "boost/1.75.0"
     build_requires = "catch2/2.13.4"
@@ -17,7 +18,9 @@ class CoLibConan(ConanFile):
     }
 
     def package(self):
-        self.copy("*.hpp")
+        self.copy("*.hpp", dst="include", src="include")
+        self.copy("*.lib", dst="lib", keep_path=False)
+        self.copy("*.a", dst="lib", keep_path=False)
 
     def build(self): # this is not building a library, just tests
         cmake = CMake(self)
@@ -25,5 +28,5 @@ class CoLibConan(ConanFile):
         cmake.build()
         cmake.test()
 
-    def package_id(self):
-        self.info.header_only()
+    def package_info(self):
+        self.cpp_info.libs = ["co_lib"]
