+++ include/co/co.hpp
@@ -2,6 +2,7 @@
 
 #include <co/channel.hpp>
 #include <co/condition_variable.hpp>
+#include <co/event.hpp>
 #include <co/func.hpp>
 #include <co/future.hpp>
 #include <co/mutex.hpp>
