+++ include/co/event.hpp
@@ -1,11 +1,16 @@
 #pragma once
 
+#include <co/error_code.hpp>
 #include <co/impl/thread_storage.hpp>
-#include <co/scheduler.hpp>
+#include <co/result.hpp>
 #include <co/std.hpp>
+#include <co/stop_token.hpp>
+#include <co/until.hpp>
 
 namespace co
 {
+
+class event;
 
 namespace impl
 {
@@ -19,7 +24,6 @@ enum class event_status
     timeout
 };
 
-template <typename event>
 class event_awaiter
 {
 public:
@@ -39,9 +43,43 @@ private:
     event& _event;
 };
 
+class interruptible_event_awaiter
+{
+public:
+    explicit interruptible_event_awaiter(event& event)
+        : interruptible_event_awaiter(event, std::nullopt, std::nullopt)
+    {}
+
+    interruptible_event_awaiter(event& event, const co::until& until)
+        : interruptible_event_awaiter(event, until.milliseconds(), until.token())
+    {}
+
+    interruptible_event_awaiter(event& event,
+                                std::optional<int64_t> milliseconds,
+                                const std::optional<co::stop_token>& tokenOpt);
+
+    interruptible_event_awaiter& operator=(const interruptible_event_awaiter&) = delete;
+    interruptible_event_awaiter& operator=(interruptible_event_awaiter&&) = delete;
+    interruptible_event_awaiter(interruptible_event_awaiter&&) = delete;
+    interruptible_event_awaiter(const interruptible_event_awaiter&) = delete;
+
+    bool await_ready() const noexcept;
+    void await_suspend(std::coroutine_handle<> awaiting_coroutine) noexcept;
+    result<void> await_resume();
+
+private:
+    co::stop_callback_func stop_callback_func();
+    static void on_timer(void* awaiter_ptr);
+
+    thread_storage* _thread_storage = nullptr;  // the thread to which the awaiter belongs
+    event& _event;
+    const std::optional<int64_t> _milliseconds_opt;
+    std::optional<co::stop_callback> _stop_callback;
+};
+
 }  // namespace impl
 
-/// \brief the simplest non interruptable awaitable object
+/// \brief is a interruptable synchronisation primitive for one time notification
 ///
 /// Example:
 /// \code
@@ -50,38 +88,34 @@ private:
 ///     {
 ///         e.notify();
 ///     }).detach();
-///     co_await e.wait();
+///     auto res = co_await e.wait(co::until(100ms, co::this_thread::stop_token()));
+///     if (res == co::cancel)
+///         std::cout << "cancelled by stop token\n";
+///     else if (res == co::timeout)
+///          std::cout << "timeout\n";
+///     else
+///     {
+///         assert(res.is_ok());
+///         std::cout << "successfully notified\n";
+///     }
 /// \endcode
 class event
 {
-    template <typename T>
     friend class impl::event_awaiter;
+    friend class impl::interruptible_event_awaiter;
 
 public:
     /// \brief notify the awaited side that the event is ready.
     ///
     /// Notify can be called many times, but only first time has an effect.
     /// \return true only in case when the notification is successful. That means that awaited side will not block on
-    /// wait() or will be resumed
-    bool notify() noexcept
-    {
-        if (_status == impl::event_status::ok)
-            return false;
-
-        if (_status == impl::event_status::waiting)
-        {
-            assert(_waiting_coro != nullptr);
-            impl::get_scheduler().ready(_waiting_coro);
-        }
-
-        _status = impl::event_status::ok;
-        return true;
-    }
+    /// wait() or will be resumed without errors (co::result::is_ok() == true);
+    bool notify() noexcept;
 
     /// \brief will suspend the co::thread until the notification will be received with notify() method
     ///
     /// wait() won't block if notify() is called in advance. wait() can be called twice, however the second time it will
-    /// return immediately. The event can be waited only from one thread.
+    /// return immediately. The event can be waited only from one thread
     ///
     /// Usage:
     /// \code
@@ -89,16 +123,30 @@ public:
     /// \endcode
     ///
     /// \return void
-    [[nodiscard("co_await me")]] impl::event_awaiter<event> wait()
-    {
-        if (_status == impl::event_status::waiting)
-            throw std::logic_error("event already waiting");
+    [[nodiscard("co_await me")]] impl::event_awaiter wait();
+    ;
 
-        return impl::event_awaiter<event>(*this);
-    };
+    /// \brief will suspend the co::thread until the notification will be received with notify() method or interruption
+    /// is happened according to the until object
+    ///
+    /// wait() won't block if notify() is called in advance. wait() can be called twice, however the second time it will
+    /// return immediately. The event can be waited only from one thread
+    ///
+    /// Usage:
+    /// \code
+    ///     co::result<void> res = co_await event.wait(co::until(100ms, co::this_thread::stop_token()));
+    ///     if (res.is_err())
+    ///         std::cout << "interrupted\n";
+    /// \endcode
+    ///
+    /// \return void
+    [[nodiscard("co_await me")]] impl::interruptible_event_awaiter wait(const co::until& until);
+    ;
 
-    /// \brief check whether notify() was called
-    [[nodiscard]] bool is_notified() const
+    /// \brief check whether notify() was successfully called
+    ///
+    /// is_notified() returns false if the event was interrupted or not notified
+    bool is_notified() const
     {
         return _status == impl::event_status::ok;
     }
@@ -107,55 +155,5 @@ private:
     impl::event_status _status = impl::event_status::init;
     std::coroutine_handle<> _waiting_coro;
 };
-
-namespace impl
-{
-
-template <typename event>
-event_awaiter<event>::event_awaiter(event& _event)
-    : _thread_storage(get_this_thread_storage_ptr())
-    , _event(_event)
-{
-    // we can't run async code outside of co::thread. Then _thread_storage should be
-    // defined in any point of time
-    assert(_thread_storage != nullptr);
-}
-
-template <typename event>
-bool event_awaiter<event>::await_ready() const noexcept
-{
-    return (_event._status == event_status::ok);
-}
-
-template <typename event>
-void event_awaiter<event>::await_suspend(std::coroutine_handle<> awaiting_coroutine) noexcept
-{
-    _event._waiting_coro = awaiting_coroutine;
-    _event._status = event_status::waiting;
-    set_this_thread_storage_ptr(nullptr);
-}
-
-template <typename event>
-void event_awaiter<event>::await_resume()
-{
-    assert(_event._status == event_status::ok);
-
-    set_this_thread_storage_ptr(_thread_storage);
-
-    switch (_event._status)
-    {
-    case event_status::init:
-    case event_status::waiting:
-    case event_status::cancel:
-    case event_status::timeout:
-        assert(false);
-        throw std::logic_error("unexpected status in event_awaiter::await_resume()");
-    case event_status::ok:
-        return;
-    }
-    assert(false);  // unreachable
-}
-
-}  // namespace impl
 
 }  // namespace co
