+++ include/co/impl/timer.hpp
@@ -19,57 +19,17 @@ public:
     timer& operator=(const timer&) = delete;
     timer& operator=(timer&&) = delete;
 
-    void init(uv_loop_t* uv_loop)
-    {
-        assert(!_uv_timer_opt.has_value());
+    void init(uv_loop_t* uv_loop);
 
-        _uv_timer_opt = uv_timer_t{};
-        int res = uv_timer_init(uv_loop, &(*_uv_timer_opt));
-        using namespace std::string_literals;
-        if (res != 0)
-            throw std::runtime_error("unable to init a timer:"s + uv_strerror(res));
-        _uv_timer_opt->data = static_cast<void*>(this);
-    }
-
-    void set_timer(int64_t milliseconds, callback cb, void* data)
-    {
-        assert(_uv_timer_opt.has_value());
-        assert(_cb == nullptr);
-        assert(_data == nullptr);
-
-        _cb = cb;
-        _data = data;
-        int res = uv_timer_start(&(*_uv_timer_opt), on_timer, milliseconds, 0);
-        using namespace std::string_literals;
-        if (res != 0)
-            throw std::runtime_error("can't set a timer:"s + uv_strerror(res));
-    }
+    void set_timer(int64_t milliseconds, callback cb, void* data);
 
     // NOTE: idempotent
-    void stop()
-    {
-        assert(_uv_timer_opt.has_value());
-        uv_timer_stop(&(*_uv_timer_opt));
-        _cb = nullptr;
-        _data = nullptr;
-    }
+    void stop();
 
     co::func<void> close();
 
 private:
-    static void on_timer(uv_timer_t* uv_timer)
-    {
-        assert(uv_timer != nullptr);
-        assert(uv_timer->data != nullptr);
-
-        auto& self = *static_cast<timer*>(uv_timer->data);
-        assert(self._cb != nullptr);
-        assert(self._data != nullptr);
-
-        self._cb(self._data);
-        self._cb = nullptr;
-        self._data = nullptr;
-    }
+    static void on_timer(uv_timer_t* uv_timer);
 
 private:
     std::optional<uv_timer_t> _uv_timer_opt;
