+++ include/co/thread.hpp
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <iostream>
+#include <co/event.hpp>
 #include <co/func.hpp>
 #include <co/impl/shared_state.hpp>
 #include <co/impl/thread_storage.hpp>
 #include <co/std.hpp>
-#include <co/timed_event.hpp>
+#include <co/impl/timer.hpp>
 
 namespace co
 {
@@ -52,7 +53,7 @@ inline thread_func thread_func_promise::get_return_object() noexcept
 }
 
 inline thread_func create_thread_main_func(func<void> func,
-                                           std::shared_ptr<timed_event> finish,
+                                           std::shared_ptr<event> finish,
                                            std::shared_ptr<thread_storage> thread_storage)
 {
     try
@@ -83,7 +84,7 @@ public:
 
     explicit thread(func<void>&& func, const std::string& thread_name = "")
         : _thread_storage_ptr(impl::create_thread_storage(thread_name, ++id))
-        , _event_ptr(std::make_shared<timed_event>())
+        , _event_ptr(std::make_shared<event>())
         , _thread_func(impl::create_thread_main_func(std::move(func), _event_ptr, _thread_storage_ptr))
     {
         // schedule the thread execution
@@ -139,7 +140,7 @@ private:
 
     bool _detached = false;
     std::shared_ptr<impl::thread_storage> _thread_storage_ptr;
-    std::shared_ptr<timed_event> _event_ptr;
+    std::shared_ptr<event> _event_ptr;
     impl::thread_func _thread_func;
 };
